Get rid of `AbstractInControl`
Use `InControl` instead.
