+++ src/containers/group.control.spec.ts
@@ -3,10 +3,10 @@ import { afterSupplied, onceAfter, onSupplied } from '@proc7ts/fun-events';
 import { InAspect__symbol } from '../aspect';
 import { InRenderScheduler } from '../aspects';
 import { InControl } from '../control';
-import { inValue } from '../controls';
 import { intoFallback } from '../conversion';
 import { InData, InMode } from '../data';
 import { InStyledElement } from '../elements/style';
+import { inValue } from '../value.control';
 import { InContainer } from './container.control';
 import { inGroup, InGroup } from './group.control';
 import { InList } from './list.control';
