+++ src/containers/list.control.spec.ts
@@ -4,8 +4,8 @@ import { Supply } from '@proc7ts/primitives';
 import { InAspect__symbol } from '../aspect';
 import { InRenderScheduler } from '../aspects';
 import { InControl } from '../control';
-import { inValue } from '../controls';
 import { InData, InMode } from '../data';
+import { inValue } from '../value.control';
 import { InContainer } from './container.control';
 import { InGroup } from './group.control';
 import { inList, InList } from './list.control';
