--- src/controls/index.ts
@@ -1,3 +1,0 @@
-export * from './abstract.control';
-export * from './value.control';
-export * from './value-of.control';
