+++ src/elements/forms/submit-button.control.spec.ts
@@ -1,9 +1,9 @@
 import { afterThe } from '@proc7ts/fun-events';
 import { asis, noop } from '@proc7ts/primitives';
 import { InControl } from '../../control';
-import { inValue } from '../../controls';
 import { InMode } from '../../data';
 import { InSubmit, InSubmitRejectedError } from '../../submit.aspect';
+import { inValue } from '../../value.control';
 import { inSubmitButton, InSubmitButton } from './submit-button.control';
 
 describe('InSubmitButton', () => {
