+++ src/index.ts
@@ -6,7 +6,6 @@ export * from './aspect';
 export * from './aspects';
 export * from './containers';
 export * from './control';
-export * from './controls';
 export * from './conversion';
 export * from './converter';
 export * from './data';
@@ -18,3 +17,4 @@ export * from './elements/forms';
 export * from './elements/style';
 export * from './submit.aspect';
 export * from './validation';
+export * from './value.control';
