+++ src/validation/require-length.validator.spec.ts
@@ -1,5 +1,5 @@
 import { InControl } from '../control';
-import { inValue } from '../controls';
+import { inValue } from '../value.control';
 import { requireLength } from './require-length.validator';
 import { InValidation } from './validation.aspect';
